[LBSE] Suppress non-foreground paint phases for SVG shapes & images https://bugs.webkit.org/show_bug.cgi?id=236189
Reviewed by NOBODY (OOPS!).

Suppress all paint phases except PaintPhase::Foreground for SVG primitives
such as paths, rects, images, etc. - all but RenderSVGContainer. We still
have to propagate the paint phases down the SVG render tree through the
container hierachy, as there could be <foreignObject> elements that do
need all paint phases.

If this ever turns out as bottleneck we can keep track of the presence
of <foreignObject> elements in a SVG subtree and only selectively propagate
the paint phases to those leafs that need them.

This optimization was used and tested with all SVG pixel tests in
the downstream branch -- so it's well tested that this works fine.

Covered by existing tests, no change in behaviour.

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintForegroundForFragments):